# Remove one kernel in per_tensor_quant_mla_fp8