--- a/python/sglang/srt/utils.py
+++ b/python/sglang/srt/utils.py
@@ -922,10 +922,10 @@ def get_zmq_socket(
     context: zmq.Context, socket_type: zmq.SocketType, endpoint: str, bind: bool
 ):
     mem = psutil.virtual_memory()
-    total_mem = mem.total / 1024**3
-    available_mem = mem.available / 1024**3
+    total_mem = mem.total / 1024 ** 3
+    available_mem = mem.available / 1024 ** 3
     if total_mem > 32 and available_mem > 16:
-        buf_size = int(0.5 * 1024**3)
+        buf_size = int(0.5 * 1024 ** 3)
     else:
         buf_size = -1
 
@@ -1445,10 +1445,10 @@ def dataclass_to_string_truncated(
         return (
             "{"
             + ", ".join(
-                f"'{k}': {dataclass_to_string_truncated(v, max_length)}"
-                for k, v in data.items()
-                if k not in skip_names
-            )
+            f"'{k}': {dataclass_to_string_truncated(v, max_length)}"
+            for k, v in data.items()
+            if k not in skip_names
+        )
             + "}"
         )
     elif dataclasses.is_dataclass(data):
@@ -1456,10 +1456,10 @@ def dataclass_to_string_truncated(
         return (
             f"{data.__class__.__name__}("
             + ", ".join(
-                f"{f.name}={dataclass_to_string_truncated(getattr(data, f.name), max_length)}"
-                for f in fields
-                if f.name not in skip_names
-            )
+            f"{f.name}={dataclass_to_string_truncated(getattr(data, f.name), max_length)}"
+            for f in fields
+            if f.name not in skip_names
+        )
             + ")"
         )
     else:
@@ -1678,7 +1678,7 @@ def configure_ipv6(dist_init_addr):
     port_str = None
     if len(addr) > end + 1:
         if addr[end + 1] == ":":
-            port_str = addr[end + 2 :]
+            port_str = addr[end + 2:]
         else:
             raise ValueError("received IPv6 address format: expected ':' after ']'")
 
@@ -1816,7 +1816,7 @@ def retry(
             if not should_retry(e):
                 raise Exception(f"retry() observe errors that should not be retried.")
 
-            delay = min(initial_delay * (2**try_index), max_delay) * (
+            delay = min(initial_delay * (2 ** try_index), max_delay) * (
                 0.75 + 0.25 * random.random()
             )
 
@@ -1933,3 +1933,10 @@ def is_fa3_default_architecture(hf_config):
     }
     return architectures[0] in default_archs
 
+
+class BumpAllocator:
+    def __init__(self):
+        self._buffer = TODO
+
+    def allocate(self):
+        return TODO