<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Thanks @Alcanderian for discussing it is acceptable to change APIs in the caller site

### Accuracy

```
python -m sglang.launch_server --model-path /dev/shm/DeepSeek-V3-0324 --trust-remote-code --tp 16 --dp 16 --enable-dp-attention --enable-deepep-moe --deepep-mode normal --disable-cuda-graph --disable-radix-cache --disable-overlap-schedule --decode-log-interval 1 --host 0.0.0.0 --port 20000 --dist-init-addr 10.10.38.8:15000 --nnodes 2 --node-rank 0
(cd /host_home/primary_synced/sglang && while true; do python3 benchmark/gsm8k/bench_sglang.py --port 20000 --parallel 1400 --num-questions 1400; done)
```

PR: 93.3

(baseline: roughly 93 when I tested in https://github.com/sgl-project/sglang/pull/5295)

Another acc test

```
(cd /host_home/primary_synced/sglang && python3 test/srt/test_mla_deepseek_v3.py)
```

baseline (#5370): 0.64~0.67

PR: 0.645

### Speed

Using command in https://github.com/sgl-project/sglang/pull/5370:

```
python3 -m sglang.bench_one_batch --model RedHatAI/DeepSeek-Coder-V2-Lite-Instruct-FP8 --batch-size 1 --input-len 128 --output-len 128 --trust-remote-code
```

baseline

```
Benchmark ...
Prefill. latency: 0.04585 s, throughput:   2791.45 token/s
Decode.  latency: 0.00547 s, throughput:    182.68 token/s
Decode.  latency: 0.00540 s, throughput:    185.02 token/s
Decode.  latency: 0.00539 s, throughput:    185.43 token/s
Decode.  latency: 0.00538 s, throughput:    185.93 token/s
Decode.  latency: 0.00535 s, throughput:    186.90 token/s
Decode.  median latency: 0.00529 s, median throughput:    188.94 token/s
```

PR

```
Benchmark ...
Prefill. latency: 0.05106 s, throughput:   2506.72 token/s
Decode.  latency: 0.00543 s, throughput:    184.17 token/s
Decode.  latency: 0.00534 s, throughput:    187.36 token/s
Decode.  latency: 0.00531 s, throughput:    188.17 token/s
Decode.  latency: 0.00530 s, throughput:    188.84 token/s
Decode.  latency: 0.00529 s, throughput:    189.15 token/s
Decode.  median latency: 0.00522 s, median throughput:    191.53 token/s
```

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.